Conflicts:
	CMakeLists.txt
	cmake/ConfigureSummary.cmake
	nestkernel/CMakeLists.txt
	nestkernel/connection_manager.cpp
	nestkernel/connection_manager.h
	nestkernel/exceptions.h
	nestkernel/sonata_connector.cpp
	nestkernel/sonata_connector.h
	pynest/CMakeLists.txt
	pynest/examples/aeif_cond_beta_multisynapse.py
	pynest/examples/sonata_network.py
	pynest/nest/lib/hl_api_connections.py
	pynest/nest/lib/hl_api_nodes.py
	pynest/nest/lib/hl_api_sonata.py
	pynest/nestkernel_api.pxd
	pynest/nestkernel_api.pyx

+++ models/spike_recorder.cpp
@@ -95,7 +95,7 @@ nest::spike_recorder::handle( SpikeEvent& e )
   {
     assert( e.get_multiplicity() > 0 );
 
-    for ( int i = 0; i < e.get_multiplicity(); ++i )
+    for ( size_t i = 0; i < e.get_multiplicity(); ++i )
     {
       write( e, RecordingBackend::NO_DOUBLE_VALUES, RecordingBackend::NO_LONG_VALUES );
     }
